week6/QueueProcessorProj/requirements.txt
azure-functions==1.23.0
MarkupSafe==3.0.2
Werkzeug==3.1.3
torch==2.1.2
